MVP of PostFeed

+++ backend/proj2/src/main/java/com/example/proj2/Repository/PostRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.example.proj2.Entity.Post;
 
 @Repository
-public interface PostRepository extends JpaRepository<Post, Integer>
+public interface PostRepository extends JpaRepository<Post, Long>
 {
